Caching system for CAD output
Especially for large scenarios (e.g., Outdoor1), reading the CAD file can be the bottleneck of the simulator. Since most likely the user will have a number of scenarios that will be reused over and over, caching the CAD pre-processing trades memory for speed.
Note that the code checks whether the cache is up-to-date with the CAD file or not, in which case it recomputes it and saves it. CAD cache is not stored on GitHub as it might be significantly large.