Improve performance of StringOutputGenerator
String concatention is a heavy operation. Minimized the number of successive concatenations in the function and improved code reusability and readability.
The function could be further improved: num2str is suboptimal (especially when iteratively concatenating its output).
It has not been done since functions such as sprintf return different precision with respect to num2str, making the output slightly different and thus making the code fail the tests.
Also, instead of first creating a long string, the output could be directly written to file. Using 'A' instead of 'w' prevents automatic buffer flushing and allows to append to existing file.