Halving Tx/Rx couples
Innermost for loop (over iterateRx) now starts from iterateTx+1 instead of 1, effectively halving the number of pairs of nodes to ray-trace.
Since both TxiRxj.txt and TxjRxi.txt are requested as output, a function swapping Rx and Tx output columns is created.
NOTE: polarization not well supported by the swapping function
Since computing the reverse path and simply swapping the columns may change the order of the different rays, readQdFile adds an input allowing you to sort the struct on a time-step basis. More info can be read in the documentation of the function.
This was necessary for the unit testing to not fail.