Writing QdFiles with fprintf and 'At' permission
To optimize writing speed to file, fprintf is used with "append and don't flush" permission. This allows to gain a big performance improvement.
Also, qdFilesFloatPrecision has been added to decide the output precision as '%.(qdFilesFloatPrecision)g'.
The slight change in how the QdFiles are written requires an update of such files in order for the tests to not fail.